Remove messages from Thread entity
Even though they seem connected, they are in fact not. Thread never
loads all its messages, so there's no point in modelling this relation
in it. Instead, we just need a message count to be able to display. This
can be achieved in a much cost-effective way, without instantiating
objects for each message.