Fix seeds
After recent changes seeds were broken. I did not notice it until I had
to reinstall the database. With this commit I rewrote the seeds to
a) work
b) be a class, not a raw script
c) have a sanity test that they work, so I don't accidentally break them
   next time. Unfortunately, this test is super-slow compared to other
   test, so I have to think about different way to verify them, perhaps
   by running on the CI.