Improve current_user handling across slices
Now every slice is responsible for its own handling of the current user.
It should fetch it and return a slice-specific entity, representing it.
They should be able to be different between the slices, because they are
different contexts.

This also allows me to delete some questionable choices from the past,
such as using a pseudo anti-corruption layer for no good reason (is
wasn't interacting with a legacy code or anything). The resulting code
should be easier to reason about.

This is also a preparation work for server-side sessions. The top-level
Palaver::Action is now responsible only for reading current_user_id from
the session and put it inside the response object. In the future,
session will include only a session_id and the top-level action would
read it, call the database to translate the session_id to an account id
and put the account_id in the response object. Then the slices will read
it and proceed as they do after this commit.