Replace query object with queries
Previous implementation with one class per query was a bit too much. It
was also heavily relying on the data exposed from the repository, making
it a very leaky abstraction.

This PR introduces instead a concept of query classes which are driven
by a specific querying needs and, instead of domain entities, return
just raw-ish data (in for of Ruby's `Data` objects). These queries
object use relations directly, skipping repositories completely and
letting them focus on a real domain work.

This makes this quering classes on the same level as repositories,
instead of pretending to be a level higher, while in fact it was a
convoluted repository.